refactor(app): App 退化为组合根，启动/布局/提示等拆成 app 级 hooks
App.tsx（465→217 行）只剩「hook 调用 + 少量派生 + 渲染」。抽出 hooks/：
- useBootstrap   启动加载 / 语言 / poll·focus 刷新 / 向导完成 / 连接热生效（持 boot）
- usePanelLayout 左右两栏宽度+折叠+持久化
- useUpdateNotice 版本更新提示（水合/订阅/调试）
- useToast       操作级 toast + 自动消失
- useAppStores   三个全局 store 接线
- useExternalLinkGuard 全局外链跳转防护
usePullRequests 去掉 boot 耦合（selectedConn 派生回 App、focus 刷新移入 useBootstrap）。
无逻辑/界面变更。

CHANGELOG 前端重构条目改为「标题 + 嵌套子列表」呈现重点。
