+++ apps/desktop/src/renderer/src/App.tsx
@@ -1,297 +1,71 @@
-import { useCallback, useEffect, useState } from 'react';
+import { useState } from 'react';
 import { useTranslation } from 'react-i18next';
-import i18n, { resolveUiLanguage, persistLanguage } from './i18n';
-import type {
-  AppInfo,
-  AppPaths,
-  Config,
-  ConnectionSummary,
-  PrAgentStatus,
-  PrDiscoveryFilter,
-  UpdateCheckResult,
-} from '@meebox/shared';
-import { invoke, subscribe } from './api';
-import { ChatPane, CHAT_MAX_WIDTH, CHAT_MIN_WIDTH } from './components/features/chat';
-import { wireChatRunStore } from './stores/chat-run-store';
-import { wireDraftsStore } from './stores/drafts-store';
-import { wireRepoSyncStore } from './stores/repo-sync-store';
+import type { PrDiscoveryFilter } from '@meebox/shared';
+import { invoke } from './api';
+import { ChatPane } from './components/features/chat';
 import { MainPane } from './components/layout/MainPane';
 import { PrPanel } from './components/features/pr/PrPanel';
 import { PrEmpty } from './components/features/pr/PrEmpty';
-import { usePullRequests } from './components/features/pr/hooks/usePullRequests';
-import { OnboardingWizard, type OnboardingResult } from './components/features/onboarding/OnboardingWizard';
+import { OnboardingWizard } from './components/features/onboarding/OnboardingWizard';
 import { SettingsModal } from './components/features/settings/SettingsModal';
-import { Sidebar, SIDEBAR_MAX_WIDTH, SIDEBAR_MIN_WIDTH } from './components/layout/Sidebar';
+import { Sidebar } from './components/layout/Sidebar';
 import { StatusBar } from './components/layout/StatusBar';
 import { TitleBar } from './components/layout/TitleBar';
-
-interface BootstrapState {
-  info: AppInfo;
-  paths: AppPaths;
-  config: Config;
-  prAgent: PrAgentStatus;
-  connections: ConnectionSummary[];
-  lastSyncAt: string | null;
-}
+import { useToast } from './hooks/useToast';
+import { usePullRequests } from './components/features/pr/hooks/usePullRequests';
+import { useBootstrap } from './hooks/useBootstrap';
+import { usePanelLayout } from './hooks/usePanelLayout';
+import { useUpdateNotice } from './hooks/useUpdateNotice';
+import { useAppStores } from './hooks/useAppStores';
+import { useExternalLinkGuard } from './hooks/useExternalLinkGuard';
 
 export default function App() {
   const { t } = useTranslation();
-  const [boot, setBoot] = useState<BootstrapState | null>(null);
-  const [showSettings, setShowSettings] = useState(false);
-  const [fatalError, setFatalError] = useState<string | null>(null);
-  // 启动检测到的新版本（main 推 app:updateAvailable）；StatusBar 据此提示跳转下载。
-  const [updateInfo, setUpdateInfo] = useState<UpdateCheckResult | null>(null);
-  // 操作级 toast（审批 / 合并等远端动作失败时提示，区别于 fatalError 整屏报错）。
-  // key 用随机数：同样文案连续触发也能重置自动消失计时器。
-  const [toast, setToast] = useState<{ text: string; key: number } | null>(null);
-  const notifyError = useCallback((text: string): void => {
-    setToast({ text, key: Math.random() });
-  }, []);
-  // PR 列表 / 选中 / 审批 / 合并 / 刷新归 features/pr 的 usePullRequests；App 仅注入连接摘要与
-  // 就绪态，并在 bootstrap / 向导完成时经 setPrs 注入初始列表、poll tick 时调 reloadPrs。
+  const { toast, notifyError, dismiss: dismissToast } = useToast();
+  // PR 列表 / 选中 / 审批 / 合并 / 刷新 —— 领域逻辑归 usePullRequests
   const {
     prs,
     setPrs,
     selectedId,
     setSelectedId,
     selected,
-    selectedConn,
     refreshing,
     merging,
     reloadPrs,
     triggerRefresh,
     setSelectedPrStatus,
     mergeSelectedPr,
-  } = usePullRequests({ connections: boot?.connections, ready: !!boot, notifyError });
-  // toast 自动消失（6s）；key 变化即重置计时
-  useEffect(() => {
-    if (!toast) return;
-    const id = setTimeout(() => setToast(null), 6000);
-    return () => clearTimeout(id);
-    // 仅依赖 key：同一 toast 重渲不重置计时，新 toast (key 变) 才重置
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [toast?.key]);
-  // 仅调试用：localStorage 里 meebox.forceOnboarding='1' 时强制进首启向导，
-  // 不必动 config.yaml。DevTools 设值后刷新进入；走完向导会自动清掉该 flag。
-  // 详见 docs/development.md。
-  const [forceOnboarding, setForceOnboarding] = useState(
-    () => localStorage.getItem('meebox.forceOnboarding') === '1',
-  );
+  } = usePullRequests({ notifyError });
+  // 应用启动 / 全局生命周期（boot 加载、语言、poll / focus 刷新、向导完成、连接热生效）
+  const { boot, fatalError, lastSyncAt, needsOnboarding, completeOnboarding, refreshBootAndPrs, patchConfig } =
+    useBootstrap({ setPrs, reloadPrs });
+  // 布局态（左右两栏宽度 / 折叠）、版本更新提示、store 接线、外链防护——各自成 app 级 hook
+  const {
+    sidebarWidth,
+    setSidebarWidth,
+    sidebarCollapsed,
+    setSidebarCollapsed,
+    chatWidth,
+    setChatWidth,
+    chatCollapsed,
+    setChatCollapsed,
+  } = usePanelLayout();
+  const updateInfo = useUpdateNotice();
+  useAppStores();
+  useExternalLinkGuard();
+
+  const [showSettings, setShowSettings] = useState(false);
   /**
-   * M4 跨组件跳转：ChatPane finding card 点"编辑" → 这里 set →
-   * MainPane 切 tab='diff' + 透传给 DiffView → DiffView 消费完调 onConsumed 清空。
-   * 一次性 token；非 null 时 DiffView 应该 scroll + highlight (+ open edit zone
-   * 如果带 runId/findingId 能反查到 finding-source 草稿)。
-   *
-   * runId/findingId 可选：
-   * - ChatPane finding card 跳转 → 必带，DiffView 据此找草稿自动 enter edit
-   * - PublishReviewModal anchor 点击 → 只带 anchor，DiffView 仅 navigate 不进 edit
-   *   (用户在 modal 里看到某条想确认上下文，跳过去看一眼，不一定要改)
+   * M4 跨组件跳转：ChatPane finding card 点"编辑" / PublishReviewModal anchor 点击 → 这里 set →
+   * PrPanel 切到 Diff tab + 透传给 DiffView 做 scroll/highlight/(可选)open edit zone，消费完清空。
    */
   const [pendingDiffNav, setPendingDiffNav] = useState<{
     runId?: string;
     findingId?: string;
     anchor: { path: string; startLine: number; endLine: number };
   } | null>(null);
-  const [lastSyncAt, setLastSyncAt] = useState<string | null>(null);
-  // GitHub 发现分类（运行时筛选，不持久化）；仅 GitHub 活动连接时在 PR 列表展示。
+  // GitHub 发现分类（运行时筛选，不持久化）；仅活动连接支持时在 PR 列表展示。
   const [discoveryFilter, setDiscoveryFilter] = useState<PrDiscoveryFilter>('review-requested');
-  const [sidebarWidth, setSidebarWidth] = useState<number>(() => {
-    const raw = localStorage.getItem('meebox.sidebarWidth');
-    const n = raw ? Number(raw) : 360;
-    return Math.min(SIDEBAR_MAX_WIDTH, Math.max(SIDEBAR_MIN_WIDTH, Number.isFinite(n) ? n : 360));
-  });
-  const [sidebarCollapsed, setSidebarCollapsed] = useState<boolean>(
-    () => localStorage.getItem('meebox.sidebarCollapsed') === '1',
-  );
-  const [chatWidth, setChatWidth] = useState<number>(() => {
-    const raw = localStorage.getItem('meebox.chatWidth');
-    const n = raw ? Number(raw) : 360;
-    return Math.min(CHAT_MAX_WIDTH, Math.max(CHAT_MIN_WIDTH, Number.isFinite(n) ? n : 360));
-  });
-  const [chatCollapsed, setChatCollapsed] = useState<boolean>(
-    // 默认收起：M3 之前 chat 还是空壳，避免空占地方
-    () => (localStorage.getItem('meebox.chatCollapsed') ?? '1') === '1',
-  );
-  useEffect(() => {
-    localStorage.setItem('meebox.sidebarWidth', String(sidebarWidth));
-  }, [sidebarWidth]);
-  useEffect(() => {
-    localStorage.setItem('meebox.sidebarCollapsed', sidebarCollapsed ? '1' : '0');
-  }, [sidebarCollapsed]);
-  useEffect(() => {
-    localStorage.setItem('meebox.chatWidth', String(chatWidth));
-  }, [chatWidth]);
-  useEffect(() => {
-    localStorage.setItem('meebox.chatCollapsed', chatCollapsed ? '1' : '0');
-  }, [chatCollapsed]);
-
-  // 连接改动（尤其切换活动连接）后整体刷新 boot：活动连接变化后 main 端 app:connections /
-  // prs:list 都随之变，必须重拉，否则 boot.connections、PR 列表会过期。
-  const refreshBootAndPrs = useCallback(async (): Promise<void> => {
-    const [config, connections, freshPrs, lastSync] = await Promise.all([
-      invoke('config:read', undefined),
-      invoke('app:connections', undefined),
-      invoke('prs:list', undefined),
-      invoke('prs:lastSync', undefined),
-    ]);
-    setBoot((b) => (b ? { ...b, config, connections, lastSyncAt: lastSync.at } : b));
-    setPrs(freshPrs);
-    setLastSyncAt(lastSync.at);
-  }, [setPrs]);
-
-  useEffect(() => {
-    void (async () => {
-      try {
-        if (!window.api) {
-          throw new Error('preload bridge missing: window.api is undefined');
-        }
-        const [info, paths, config, prAgent, initialPrs, connections, lastSync] = await Promise.all(
-          [
-            invoke('app:info', undefined),
-            invoke('app:paths', undefined),
-            invoke('config:read', undefined),
-            invoke('app:prAgentStatus', undefined),
-            invoke('prs:list', undefined),
-            invoke('app:connections', undefined),
-            invoke('prs:lastSync', undefined),
-          ],
-        );
-        // 先按 config.language 切到目标语言并**等其资源加载完**（懒加载语言会异步拉 chunk），
-        // 再 setBoot 渲染主界面 —— 首屏直接是用户语言，不闪兜底。persist 供下次启动同步命中。
-        const lang = resolveUiLanguage(config.language);
-        persistLanguage(lang);
-        await i18n.changeLanguage(lang);
-        setBoot({ info, paths, config, prAgent, connections, lastSyncAt: lastSync.at });
-        setPrs(initialPrs);
-        setLastSyncAt(lastSync.at);
-      } catch (e) {
-        setFatalError(e instanceof Error ? e.message : String(e));
-      }
-    })();
-  }, [setPrs]);
-
-  // 运行时语言切换（如设置页改 config.language）：boot 后 language 变化即切换并回写持久化。
-  // 首次 boot 时已在上面 await 切好，这里对同值是幂等 no-op。
-  useEffect(() => {
-    if (!boot) return;
-    const lang = resolveUiLanguage(boot.config.language);
-    persistLanguage(lang);
-    void i18n.changeLanguage(lang);
-  }, [boot]);
-
-  // 启动时把 pr-agent 活动 run + 实时 stdout 流接到全局 store；ChatPane 跨 PR
-  // 切换时可读 store 拿回运行中的状态 (本组件挂载到树根，效果等价于"应用级 hook")
-  useEffect(() => wireChatRunStore(), []);
-  // 同样思路：把 repo sync 事件流接到 store，StatusBar 任意时刻可读当前活动同步任务
-  useEffect(() => wireRepoSyncStore(), []);
-  // M4 草稿事件 → store；写盘后 drafts:changed 触发指定 PR 的草稿列表自动刷新
-  useEffect(() => wireDraftsStore(), []);
-  // 版本更新：main 为单一真相源。挂载时先水合已缓存结果（设置页手动检查 / 定时检查到的新版
-  // 不会因窗口重挂载而丢失），再订阅后续广播（手动与定时检查都经此推送）。
-  useEffect(() => {
-    void invoke('app:getUpdateStatus', undefined).then((info) => {
-      if (info) setUpdateInfo(info);
-    });
-    return subscribe('app:updateAvailable', (info) => setUpdateInfo(info));
-  }, []);
-  // dev 调试钩子：控制台 dispatch CustomEvent 模拟「发现新版」以验证状态栏 chip
-  // （dev 版本通常高于 latest，自然不会触发）。detail=null 清除。
-  //   window.dispatchEvent(new CustomEvent('meebox:debug-update'))
-  //   window.dispatchEvent(new CustomEvent('meebox:debug-update', { detail: { latestVersion: '1.2.3' } }))
-  //   window.dispatchEvent(new CustomEvent('meebox:debug-update', { detail: null }))
-  useEffect(() => {
-    const onDebug = (e: Event): void => {
-      const d = (e as CustomEvent<Partial<UpdateCheckResult> | null>).detail;
-      setUpdateInfo(
-        d === null
-          ? null
-          : {
-              ok: true,
-              hasUpdate: true,
-              currentVersion: '0.0.0',
-              latestVersion: '9.9.9',
-              url: 'https://github.com/huhamhire/code-meeseeks/releases/latest',
-              ...d,
-            },
-      );
-    };
-    window.addEventListener('meebox:debug-update', onDebug);
-    return () => window.removeEventListener('meebox:debug-update', onDebug);
-  }, []);
-
-  // 全局外链跳转防护 — 所有 UGC 场景 (评论 / PR 描述 / finding / chat 等) 内
-  // 的 <a href="http(s)://"> 点击都走系统默认浏览器，不允许 Electron 在 app
-  // window 内直接跳转覆盖整个界面。capture 阶段 listener 先于 React onClick 跑
-  useEffect(() => {
-    const onClick = (e: MouseEvent): void => {
-      const target = (e.target as HTMLElement | null)?.closest?.('a[href]');
-      if (!(target instanceof HTMLAnchorElement)) return;
-      const href = target.getAttribute('href');
-      if (!href || !/^https?:\/\//.test(href)) return;
-      e.preventDefault();
-      e.stopPropagation();
-      void invoke('app:openExternal', { url: href });
-    };
-    document.addEventListener('click', onClick, true);
-    return () => document.removeEventListener('click', onClick, true);
-  }, []);
-
-  // 订阅 main 推送的 poll tick；用于刷新 statusbar "最近同步" 显示，
-  // 并顺便重拉一次 PR 列表使后台轮询新增/删除立刻反映在 UI。
-  // 同时刷新连接摘要：启动时连接的 ping（缓存 currentUser）在建窗后才完成，首轮 tick 即随其后，
-  // 借此把状态栏用户/能力位补上（否则需手动刷新才显示）。app:connections 为廉价同步调用。
-  useEffect(() => {
-    if (!window.api) return;
-    return subscribe('poll:tick', (info) => {
-      setLastSyncAt(info.at);
-      void reloadPrs();
-      void invoke('app:connections', undefined).then(
-        (connections) => {
-          setBoot((b) => (b ? { ...b, connections } : b));
-        },
-        () => {
-          /* 摘要刷新失败不影响主流程 */
-        },
-      );
-    });
-  }, [reloadPrs]);
-
-  // 首启向导完成：落盘连接（必）+ LLM / 缓存目录（按需），再重拉配置/连接/PR 更新
-  // boot。boot.config 拿到有效 active 连接后，下方 needsOnboarding 派生为 false，
-  // 向导自然卸载、切入主界面；主界面挂载后 poll:tick 订阅 + focus 刷新自然生效。
-  const completeOnboarding = useCallback(
-    async (result: OnboardingResult): Promise<void> => {
-      await invoke('config:setConnections', {
-        connections: [result.connection],
-        active_connection_id: result.connection.id,
-      });
-      if (result.llm) {
-        await invoke('config:setLlm', {
-          llm: { profiles: [result.llm], active_id: result.llm.id },
-        });
-      }
-      const trimmedRepos = result.reposDir.trim();
-      if (trimmedRepos && trimmedRepos !== (boot?.config.workspace.repos_dir ?? '')) {
-        await invoke('config:setReposDir', { reposDir: trimmedRepos });
-      }
-      const [config, connections, freshPrs, lastSync] = await Promise.all([
-        invoke('config:read', undefined),
-        invoke('app:connections', undefined),
-        invoke('prs:list', undefined),
-        invoke('prs:lastSync', undefined),
-      ]);
-      setBoot((b) => (b ? { ...b, config, connections, lastSyncAt: lastSync.at } : b));
-      setPrs(freshPrs);
-      setLastSyncAt(lastSync.at);
-      // 走完向导清掉调试 flag，避免强制模式下完成后仍被困在向导
-      if (forceOnboarding) {
-        localStorage.removeItem('meebox.forceOnboarding');
-        setForceOnboarding(false);
-      }
-    },
-    [boot, forceOnboarding, setPrs],
-  );
 
   if (fatalError) {
     return (
@@ -300,7 +74,6 @@ export default function App() {
       </div>
     );
   }
-
   if (!boot) {
     return (
       <div className="app fatal-app">
@@ -308,12 +81,6 @@ export default function App() {
       </div>
     );
   }
-
-  // gate 条件 = 有无「有效的 active 连接」：连接为空 / active 悬空都触发首启向导。
-  // 不依赖一次性 firstRun 标记 —— 用户清空连接后下次进入仍会回到向导。
-  const needsOnboarding =
-    forceOnboarding ||
-    !boot.config.connections.some((c) => c.id === boot.config.active_connection_id);
   if (needsOnboarding) {
     return (
       <OnboardingWizard
@@ -325,16 +92,19 @@ export default function App() {
     );
   }
 
+  // 选中 PR 所属连接：能力位（审批按钮降级）+ 当前 PAT 用户（判「是否自己的 PR」）。
+  const selectedConn = selected
+    ? boot.connections.find((c) => c.connectionId === selected.connectionId)
+    : undefined;
   // 有 active 连接但 LLM 未配置 → ChatPane 给出「需配置才能启用」提示并禁用输入
   const llmConfigured = boot.config.llm.profiles.some((p) => p.id === boot.config.llm.active_id);
-
   // 发现分类标签由活动连接的能力决定（GitHub 四类、Bitbucket 两类、其余无）。
   const activeConnSummary = boot.connections.find(
     (c) => c.connectionId === boot.config.active_connection_id,
   );
   const availableDiscoveryFilters = activeConnSummary?.capabilities.discoveryFilters ?? [];
   const showDiscoveryFilter = availableDiscoveryFilters.length > 0;
-  // 选中的分类可能因切换连接而对当前平台无效（如 github 的 mentioned 切到 bitbucket）→ 回落首个可用。
+  // 选中的分类可能因切换连接而对当前平台无效 → 回落首个可用。
   const effectiveDiscoveryFilter = availableDiscoveryFilters.includes(discoveryFilter)
     ? discoveryFilter
     : availableDiscoveryFilters[0];
@@ -372,9 +142,7 @@ export default function App() {
             <PrEmpty hasConnections={boot.config.connections.length > 0} />
           )}
         </MainPane>
-        {/* ChatPane 始终挂载，折叠只是 CSS 隐藏：保住运行中的 run 生命周期。
-            如果走条件渲染，折叠 = 卸载组件，进行中的计时器 / runProgress 订阅
-            全丢，再展开只能从持久化里看到已完成的结果 */}
+        {/* ChatPane 始终挂载，折叠只是 CSS 隐藏：保住运行中的 run 生命周期（计时器 / runProgress 订阅）。 */}
         <ChatPane
           pr={selected}
           prAgent={boot.prAgent}
@@ -384,11 +152,9 @@ export default function App() {
           llmConfigured={llmConfigured}
           maxConcurrency={boot.config.pr_agent.max_concurrency}
           onOpenSettings={() => setShowSettings(true)}
-          onJumpToDraftEditor={(t) => setPendingDiffNav(t)}
+          onJumpToDraftEditor={(target) => setPendingDiffNav(target)}
           onNavigateToAnchor={(anchor) => setPendingDiffNav({ anchor })}
           onSetReviewStatus={(s) => void setSelectedPrStatus(s)}
-          // 当前 active LLM profile.model — RunningView 显示成 chip 让用户知道
-          // 这次 review 用的什么模型 (不同 profile 出的结果差异大)
           currentLlmModel={
             boot.config.llm.profiles.find((p) => p.id === boot.config.llm.active_id)?.model ?? null
           }
@@ -410,7 +176,7 @@ export default function App() {
         onSwitchActiveLlm={(id) => {
           const next = { ...boot.config.llm, active_id: id };
           void invoke('config:setLlm', { llm: next });
-          setBoot((b) => (b ? { ...b, config: { ...b.config, llm: next } } : b));
+          patchConfig((c) => ({ ...c, llm: next }));
         }}
         onJumpToPr={setSelectedId}
         updateInfo={updateInfo}
@@ -418,20 +184,10 @@ export default function App() {
         onToggleAutopilot={() => {
           const enabled = !boot.config.agent.autopilot.enabled;
           void invoke('agent:setAutopilotEnabled', { enabled });
-          setBoot((b) =>
-            b
-              ? {
-                  ...b,
-                  config: {
-                    ...b.config,
-                    agent: {
-                      ...b.config.agent,
-                      autopilot: { ...b.config.agent.autopilot, enabled },
-                    },
-                  },
-                }
-              : b,
-          );
+          patchConfig((c) => ({
+            ...c,
+            agent: { ...c.agent, autopilot: { ...c.agent.autopilot, enabled } },
+          }));
         }}
       />
       {showSettings && (
@@ -439,13 +195,9 @@ export default function App() {
           info={boot.info}
           paths={boot.paths}
           config={boot.config}
-          onLlmChange={(llm) => setBoot((b) => (b ? { ...b, config: { ...b.config, llm } } : b))}
-          onProxyChange={(proxy) =>
-            setBoot((b) => (b ? { ...b, config: { ...b.config, proxy } } : b))
-          }
-          onLanguageChange={(language) =>
-            setBoot((b) => (b ? { ...b, config: { ...b.config, language } } : b))
-          }
+          onLlmChange={(llm) => patchConfig((c) => ({ ...c, llm }))}
+          onProxyChange={(proxy) => patchConfig((c) => ({ ...c, proxy }))}
+          onLanguageChange={(language) => patchConfig((c) => ({ ...c, language }))}
           onConnectionsChange={refreshBootAndPrs}
           onClose={() => setShowSettings(false)}
         />
@@ -454,7 +206,7 @@ export default function App() {
         <div
           className="app-toast app-toast-error"
           role="alert"
-          onClick={() => setToast(null)}
+          onClick={dismissToast}
           title={t('app.toastCloseTitle')}
         >
           {toast.text}
